refactor(tests): reorganize and rename test files
