+++ tests/components/AuthButtons.test.tsx
@@ -2,7 +2,7 @@ import AuthButtons from '@/components/shared/AuthButtons';
 import { render, screen } from '@testing-library/react';
 import { useSession } from 'next-auth/react';
 import { NextIntlClientProvider } from 'next-intl';
-import messages from '../messages/en.json';
+import messages from '../../messages/en.json';
 
 describe('AuthButtons', () => {
   it('renders sign in button when unauthenticated', async () => {
@@ -19,7 +19,7 @@ describe('AuthButtons', () => {
     expect(signUpButton).toBeInTheDocument();
   });
 
-  it('renders sign out button when authenticated', async () => {
+  it('renders sign out and main page buttons when authenticated', async () => {
     (useSession as jest.Mock).mockReturnValue({
       data: { user: { name: 'User', email: 'test@example.com' } },
       status: 'authenticated'
@@ -31,7 +31,10 @@ describe('AuthButtons', () => {
       </NextIntlClientProvider>
     );
 
-    const signOutButton = await screen.findByRole('button', { name: /sign out/i });
+    const mainPageButton = await screen.findByText(messages.button.mainPage);
+    expect(mainPageButton).toBeInTheDocument();
+
+    const signOutButton = await screen.findByText(messages.button.signOut);
     expect(signOutButton).toBeInTheDocument();
   });
 });
