+++ tests/components/Header.test.tsx
@@ -1,5 +1,7 @@
 import Header from '@/components/shared/layout/Header';
 import { render, screen } from '@testing-library/react';
+import messages from '../../messages/en.json';
+import { NextIntlClientProvider } from 'next-intl';
 
 jest.mock('@/components/shared/AuthButtons', () => () => <div>AuthButtonsMock</div>);
 jest.mock('@/components/shared/layout/LocaleSwitcher', () => () => <div>LocaleSwitcherMock</div>);
@@ -8,9 +10,13 @@ jest.mock('next/link', () => ({ children, href }: any) => <a href={href}>{childr
 
 describe('Header', () => {
   it('renders title and AuthButtons', () => {
-    render(<Header />);
+    render(
+      <NextIntlClientProvider locale="en" messages={messages}>
+        <Header />
+      </NextIntlClientProvider>
+    );
 
-    expect(screen.getByText('REST Client App')).toBeInTheDocument();
+    expect(screen.getByText(messages.Header.title)).toBeInTheDocument();
 
     expect(screen.getByText('AuthButtonsMock')).toBeInTheDocument();
     expect(screen.getByText('LocaleSwitcherMock')).toBeInTheDocument();
