+++ tests/components/LocaleSwitcher.test.tsx
