+++ tests/components/MainContent.test.tsx
@@ -1,6 +1,8 @@
 import { render, screen } from '@testing-library/react';
 import MainContent from '@/components/main/MainContent';
 import { useSession } from 'next-auth/react';
+import { NextIntlClientProvider } from 'next-intl';
+import messages from '../../messages/en.json';
 
 jest.mock('@/components/shared/AuthButtons', () => () => <div>AuthButtonsMock</div>);
 
@@ -8,13 +10,21 @@ describe('MainContent', () => {
   it('renders loading state', () => {
     (useSession as jest.Mock).mockReturnValue({ data: null, status: 'loading' });
 
-    render(<MainContent />);
-    expect(screen.getByText('Loading...')).toBeInTheDocument();
+    render(
+      <NextIntlClientProvider locale="en" messages={messages}>
+        <MainContent />
+      </NextIntlClientProvider>
+    );
+    expect(screen.getByText(messages.status.loading)).toBeInTheDocument();
   });
 
   it('renders unauthenticated user', () => {
     (useSession as jest.Mock).mockReturnValue({ data: null, status: 'unauthenticated' });
-    render(<MainContent />);
+    render(
+      <NextIntlClientProvider locale="en" messages={messages}>
+        <MainContent />
+      </NextIntlClientProvider>
+    );
 
     expect(screen.getByText('Welcome!')).toBeInTheDocument();
     expect(screen.getByText('AuthButtonsMock')).toBeInTheDocument();
@@ -26,7 +36,11 @@ describe('MainContent', () => {
       status: 'authenticated'
     });
 
-    render(<MainContent />);
+    render(
+      <NextIntlClientProvider locale="en" messages={messages}>
+        <MainContent />
+      </NextIntlClientProvider>
+    );
     expect(screen.getByText('Welcome, User!')).toBeInTheDocument();
     expect(screen.getByText('History')).toBeInTheDocument();
     expect(screen.queryByText('AuthButtonsMock')).not.toBeInTheDocument();
